Frontend/design - Buyer and Broker dashboard initial styling
Created placeholders for tables and charts for the following:
Started the initial styling for approval from the team :
- Buyer Home (dashboard) [WIP]
- Buyer Transaction page [Todo]
- Buyer Stock page [Todo]
- Buyer Portfolio page[Todo]
- Broker Home (dashboard) [WIP]
- Broker Transaction page [Todo]
- Broker Stock page [Todo]]
- Broker Portfolio page [Todo]

Added quality of life fixes especially to the views and added spacing to the flash messages.
Discussed to either add a timeout to the notice or add a dismiss/close button for the notice messages.

Based our initial design from https://dashboard.io/demo/ and Google Adsense Dashboard
Will implement chartkick gem to make the charts.