[WIP]Buyer and Broker pages styling (Transaction and Portfolio)
#### Buyer and Broker pages styling (Transaction and Portfolio)

## Frontend Styling using Bootstrap

Though my initial style and mockups was approved by the team.

We did more research and found a new reference for our dashboards and what we like our users to see. Aside from the Google Adsense Dashboard reference, this is a bit closer to what our team would like to achieve, once we completed the backend codes. The team will make a cohesive dashboard based on these :
https://www.google.com/finance/
https://finance.yahoo.com/

Also found a ticker widget that we implemented on the Buyer and Broker dashboard

https://www.tradingview.com/widget/ticker-tape/